# Reduce O(N) queries during startup