- avoid lookup-by-name of user and admin roles each time they are assigned
- during admin assignment, filter users-to-update to only those that need updating, which will usually be empty

No change in behavior, but dramatically fewer queries

pulled out of #4494

one of the issues raised in #4485

One thing that would be handy - emit a metric for the number of queries performed for each test. I'm not sure how to do that.