@minrk 2023-06-28
Sample startup performance with an artificial 1ms query latency for 5000 active users shows reducing startup time from 50 seconds to ~2 seconds, and (more realistically) 5000 users with few active from ~20 seconds to ~0.4 seconds.

![image](https://github.com/jupyterhub/jupyterhub/assets/151929/717ad9cc-b44f-4cdc-91bb-7c2f6137a16b)

https://gist.github.com/minrk/aba0d612b04f7603bc5688b7d8bbd187

benchmark script: https://github.com/minrk/jupyterhub-bench